Remove old mqtt service
