+++ src/app/components/AcLoads.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react"
 import NumericValue from "./NumericValue"
 import MqttSubscriptions from "../mqtt/MqttSubscriptions"
-import { phaseSum } from "../../service/util"
+import { phaseSum } from "../utils/util"
 
 const getTopics = (portalId, vebusInstanceId) => {
   return {
