--- src/service/index.js
@@ -1,3 +1,0 @@
-import VenusClient from "./VenusClient"
-
-export default VenusClient
